Produto Dao DaoJDBC and Controller modified

+++ src/Model/DAO/ProdutoDao.java
@@ -8,7 +8,7 @@ public interface ProdutoDao {
 	
 	void insert(Produto produto);
 	void update(Produto produto);
-	void deleteById(Long id);
-	Produto findById(Long id);
+	void deleteById(Integer id);
+	Produto findById(Integer id);
 	List<Produto> findAll();
 }
